+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/token/AccessTokenProvider.java
@@ -3,7 +3,7 @@ package com.barlow.core.auth.authentication.token;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.barlow.core.auth.authentication.core.MemberPrincipal;
-import com.barlow.core.support.jwt.JwtConfig;
+import com.barlow.core.auth.config.JwtConfig;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
