+++ core/core-auth/src/main/java/com/barlow/core/auth/authentication/token/AccessTokenValidator.java
@@ -7,7 +7,7 @@ import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.Claim;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.barlow.core.support.jwt.JwtConfig;
+import com.barlow.core.auth.config.JwtConfig;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
